# api/generated/openapi/aex-central.json
{
  "components": {
...
    "securitySchemes": {
      "accountToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The AEX central plane.",
    "title": "AEX bootstrap and control API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/sessions": {
      "post": {
        "operationId": "dashboard_session_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "aex:schema:DashboardSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "aex:schema:DashboardSessionCredential"
                }
              }
            },
            "description": "Success."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "aex:schema:ApiError"
                }
              }
            },
            "description": "Error."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "aex:schema:ApiError"
                }
              }
            },
            "description": "Error."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "aex:schema:ApiError"
                }
              }
            },
            "description": "Error."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "aex:schema:ApiError"
                }
              }
            },
            "description": "Error."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "aex:schema:ApiError"
                }
              }
            },
            "description": "Error."
          }
        },
        "summary": "Exchange a provider authorization code for a browser session.",
        "tags": [
          "auth"
        ],
        "x-aex-alt-principal": "anonymous",
        "x-aex-body-class": "aex_json",
        "x-aex-errors": [
          "unauthenticated",
          "forbidden",
          "invalid_request",
          "rate_limited",
          "internal_error"
        ],
        "x-aex-etag": "none",
        "x-aex-idempotency": "none",
        "x-aex-pause-exempt": true,
        "x-aex-safe-retry": false,
        "x-aex-scope": null,
        "x-aex-transport": "unary"
      }
    },
...
  },
  "security": [
    {
      "accountToken": []
    }
  ],
  "servers": [
    {
      "description": "The single bootstrap host; it is not regional.",
      "url": "https://api.aex.dev"
    }
  ]
}